refactor(eidos): extract knowledge::path submodule (#2642)
Move PathValidationLayer, PathValidationError, ValidatedPath,
validate_memory_path, and all path validation helpers/constants
to knowledge/path.rs. mod.rs is now purely re-exports and the
display_via_as_str macro.
